+++ qoffer-infrastructure/src/main/groovy/life/qbic/portal/qoffer2/web/views/create/offer/CustomerSelectionView.groovy
@@ -42,6 +42,8 @@ class CustomerSelectionView extends VerticalLayout{
     Grid<Affiliation> affiliationGrid
     HorizontalLayout affiliationLayout
     VerticalLayout affiliationLabelLayout
+    HorizontalLayout createAffiliationLayout
+    Button createAffiliationButton
 
     CustomerSelectionView(CreateOfferViewModel viewModel){
         this.viewModel = viewModel
@@ -97,7 +99,7 @@ class CustomerSelectionView extends VerticalLayout{
         this.affiliationGrid = new Grid<>()
         affiliationLayout = new HorizontalLayout(affiliationGrid)
 
-        this.addComponents(customerLayout, addButtonsLayout ,buttonLayout)
+        this.addComponents(customerLayout, addButtonsLayout , buttonLayout)
     }
 
     /**
@@ -147,6 +149,13 @@ class CustomerSelectionView extends VerticalLayout{
             //specify size of grid and layout
             affiliationLayout.setSizeFull()
             affiliationGrid.setSizeFull()
+            createAffiliationLayout = new HorizontalLayout()
+            createAffiliationButton = new Button("Create Affiliation", VaadinIcons.OFFICE)
+            createAffiliationButton.addStyleName(ValoTheme.BUTTON_FRIENDLY)
+
+            createAffiliationLayout.addComponent(createAffiliationButton)
+            createAffiliationLayout.setComponentAlignment(createAffiliationButton, Alignment.MIDDLE_RIGHT)
+            createAffiliationLayout.setSizeFull()
 
         } catch (Exception e) {
             new Exception("Unexpected exception in building the affiliation grid", e)
@@ -168,6 +177,8 @@ class CustomerSelectionView extends VerticalLayout{
 
             this.addComponent(affiliationLabelLayout,2)
             this.addComponent(affiliationGrid,3)
+            this.addComponent(createAffiliationLayout,4)
+
         })
 
         affiliationGrid.addSelectionListener({